+++ .coverage
@@ -1,1 +1,1 @@
-!coverage.py: This is a private format, don't read it directly!{"lines":{"/home/mintadmin/.ipython/profile_bluesky/startup/APS_BlueSky_tools/tests/__main__.py":[4,5,6,11,14,26,27,28,16,18,21,22,23],"/home/mintadmin/.ipython/profile_bluesky/startup/APS_BlueSky_tools/tests/test_simple.py":[4,6,7,8,10,11,12,13,15,18,20,23,26,30,40,31,33,35,36,37,21,27,24],"/home/mintadmin/.ipython/profile_bluesky/startup/APS_BlueSky_tools/APS_BlueSky_tools/__init__.py":[4,5,6,7,8,9,10,11,12,13,14,15,17,18,20,21,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,48,55,56,57]}}
+!coverage.py: This is a private format, don't read it directly!{"lines":{"/home/mintadmin/.ipython/profile_bluesky/startup/apstools/tests/__main__.py":[4,5,6,11,14,26,27,28,16,18,21,22,23],"/home/mintadmin/.ipython/profile_bluesky/startup/apstools/tests/test_simple.py":[4,6,7,8,10,11,12,13,15,18,20,23,26,30,40,31,33,35,36,37,21,27,24],"/home/mintadmin/.ipython/profile_bluesky/startup/apstools/apstools/__init__.py":[4,5,6,7,8,9,10,11,12,13,14,15,17,18,20,21,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,48,55,56,57]}}
