+++ .gitattributes
@@ -1,1 +1,1 @@
-APS_BlueSky_tools/_version.py export-subst
+apstools/_version.py export-subst
