+++ MANIFEST.in
@@ -1,5 +1,5 @@
 include README.md
 include CHANGES*
-include APS_BlueSky_tools/LICENSE
+include apstools/LICENSE
 include versioneer.py
-include APS_BlueSky_tools/_version.py
+include apstools/_version.py
