+++ README.md
@@ -1,16 +1,16 @@
-# APS_BlueSky_tools
+# apstools
 
 Various Python tools for use with BlueSky at the APS
 
-[![Python version](https://img.shields.io/pypi/pyversions/APS_BlueSky_tools.svg)](https://pypi.python.org/pypi/APS_BlueSky_tools)
+[![Python version](https://img.shields.io/pypi/pyversions/apstools.svg)](https://pypi.python.org/pypi/apstools)
 [![Documentation Status](https://readthedocs.org/projects/aps-bluesky-tools/badge/?version=latest)](http://aps-bluesky-tools.readthedocs.io/en/latest/?badge=latest)
-[![tag](https://img.shields.io/github/tag/BCDA-APS/APS_BlueSky_tools.svg)](https://github.com/BCDA-APS/APS_BlueSky_tools/tags)
-[![release](https://img.shields.io/github/release/BCDA-APS/APS_BlueSky_tools.svg)](https://github.com/BCDA-APS/APS_BlueSky_tools/releases)
-[![PyPi](https://img.shields.io/pypi/v/APS_BlueSky_tools.svg)](https://pypi.python.org/pypi/APS_BlueSky_tools)
-[![conda](https://anaconda.org/prjemian/APS_BlueSky_tools/badges/version.svg)](https://anaconda.org/prjemian/APS_BlueSky_tools)
-[![unit test](https://travis-ci.org/BCDA-APS/APS_BlueSky_tools.svg?branch=master)](https://travis-ci.org/BCDA-APS/APS_BlueSky_tools)
-[![coverage](https://coveralls.io/repos/github/BCDA-APS/APS_BlueSky_tools/badge.svg?branch=master)](https://coveralls.io/github/BCDA-APS/APS_BlueSky_tools?branch=master)
-[![code health](https://landscape.io/github/BCDA-APS/APS_BlueSky_tools/master/landscape.svg?style=plastic)](https://landscape.io/github/BCDA-APS/APS_BlueSky_tools/master)
+[![tag](https://img.shields.io/github/tag/BCDA-APS/apstools.svg)](https://github.com/BCDA-APS/apstools/tags)
+[![release](https://img.shields.io/github/release/BCDA-APS/apstools.svg)](https://github.com/BCDA-APS/apstools/releases)
+[![PyPi](https://img.shields.io/pypi/v/apstools.svg)](https://pypi.python.org/pypi/apstools)
+[![conda](https://anaconda.org/prjemian/apstools/badges/version.svg)](https://anaconda.org/prjemian/apstools)
+[![unit test](https://travis-ci.org/BCDA-APS/apstools.svg?branch=master)](https://travis-ci.org/BCDA-APS/apstools)
+[![coverage](https://coveralls.io/repos/github/BCDA-APS/apstools/badge.svg?branch=master)](https://coveralls.io/github/BCDA-APS/apstools?branch=master)
+[![code health](https://landscape.io/github/BCDA-APS/apstools/master/landscape.svg?style=plastic)](https://landscape.io/github/BCDA-APS/apstools/master)
 
 * http://nsls-ii.github.io/
 * https://github.com/NSLS-II/bluesky
@@ -22,7 +22,7 @@ item              | description
 **author**        | Pete R. Jemian
 **email**         | jemian@anl.gov
 **copyright**     | 2017-2019, Pete R. Jemian
-[**license**](APS_BlueSky_tools/LICENSE) | ANL OPEN SOURCE LICENSE
-[**documentation**](https://APS_BlueSky_tools.readthedocs.io) | https://APS_BlueSky_tools.readthedocs.io
-**source**        | https://github.com/BCDA-APS/APS_BlueSky_tools
+[**license**](apstools/LICENSE) | ANL OPEN SOURCE LICENSE
+[**documentation**](https://apstools.readthedocs.io) | https://apstools.readthedocs.io
+**source**        | https://github.com/BCDA-APS/apstools
