+++ apstools/LICENSE
@@ -1,8 +1,8 @@
-Copyright (c) 2017-2018, UChicago Argonne, LLC
+Copyright (c) 2017-2019, UChicago Argonne, LLC
 
 All Rights Reserved
 
-APS_BlueSky_tools
+apstools (previously: APS_BlueSky_tools)
 
 BCDA, Advanced Photon Source, Argonne National Laboratory
