+++ apstools/examples.py
@@ -45,7 +45,7 @@ def plan_catalog(db):
     
     Example::
     
-        from APS_BlueSky_tools.examples import plan_catalog
+        from apstools.examples import plan_catalog
         plan_catalog(db)
     
     """
@@ -74,7 +74,7 @@ def main():
     """
     summary list of all scans in the databroker
 
-    ``aps_bluesky_tools_plan_catalog`` command-line application
+    ``apstools_plan_catalog`` command-line application
     
     This can be unwieldy if there are many scans in the databroker.
     Consider it as a demo program rather than for general, long-term use.
