+++ apstools/legacy_code___devices.py
@@ -1,81 +1,55 @@
+#
+# this file will be removed before initial release!
+#
 
 """
 (ophyd) Devices that might be useful at the APS using BlueSky
-
-APS GENERAL SUPPORT
-
-.. autosummary::
-   
-    ~ApsMachineParametersDevice
-    ~ApsPssShutter
-    ~ApsPssShutterWithStatus
-    ~SimulatedApsPssShutterWithStatus
-
-AREA DETECTOR SUPPORT
 
 .. autosummary::
    
     ~AD_setup_FrameType
     ~AD_warmed_up
-    ~AD_EpicsHdf5FileName
-
-DETECTOR / SCALER SUPPORT
-
-.. autosummary::
-   
-    ~use_EPICS_scaler_channels
-
-MOTORS, POSITIONERS, AXES, ...
-
-.. autosummary::
-   
-    ~AxisTunerException
-    ~AxisTunerMixin
-    ~EpicsDescriptionMixin
-    ~EpicsMotorDialMixin
-    ~EpicsMotorLimitsMixin
-    ~EpicsMotorRawMixin
-    ~EpicsMotorServoMixin
-    ~EpicsMotorShutter
-    ~EpicsOnOffShutter
-
-SHUTTERS
-
-.. autosummary::
-   
+    ~ApsHDF5Plugin
+    ~ApsMachineParametersDevice
     ~ApsPssShutter
     ~ApsPssShutterWithStatus
+    ~AxisTunerException
+    ~AxisTunerMixin
+    ~DualPf4FilterBox
+    ~EpicsMotorDialMixin
+    ~EpicsMotorLimitsMixin
+    ~EpicsMotorServoMixin
+    ~EpicsMotorRawMixin
+    ~EpicsMotorDescriptionMixin
     ~EpicsMotorShutter
     ~EpicsOnOffShutter
-
-synApps records
-
-.. autosummary::
-   
-    ~busyRecord
-    ~sscanRecord
+    ~ProcedureRegistry
+    ~SimulatedApsPssShutterWithStatus
+    ~sscanRecord  
     ~sscanDevice
     ~swaitRecord
     ~swait_setup_random_number
     ~swait_setup_gaussian
     ~swait_setup_lorentzian
     ~swait_setup_incrementer
+    ~TunableEpicsMotor
+    ~use_EPICS_scaler_channels
     ~userCalcsDevice
-
-OTHER SUPPORT
-
-.. autosummary::
-   
-    ~DualPf4FilterBox
-    ~EpicsDescriptionMixin
-    ~ProcedureRegistry
 
 Internal routines
 
 .. autosummary::
 
     ~ApsOperatorMessagesDevice
-    ~DeviceMixinBase
+    ~ApsFileStoreHDF5
+    ~ApsFileStoreHDF5IterativeWrite
+
+Legacy routines
+
+.. autosummary::
+   
+    ~EpicsMotorWithDial
+    ~EpicsMotorWithServo
 
 """
 
@@ -97,7 +71,6 @@ import ophyd
 from ophyd import Component, Device, DeviceStatus, FormattedComponent
 from ophyd import Signal, EpicsMotor, EpicsSignal, EpicsSignalRO
 from ophyd.scaler import EpicsScaler, ScalerCH
-from ophyd.positioner import PositionerBase
 
 from ophyd.areadetector.filestore_mixins import FileStoreHDF5
 from ophyd.areadetector.filestore_mixins import FileStoreBase
@@ -105,10 +78,6 @@ from ophyd.areadetector.filestore_mixins import FileStorePluginBase
 from ophyd.areadetector.filestore_mixins import FileStoreIterativeWrite
 from ophyd import HDF5Plugin
 from ophyd.utils import set_and_wait
-
-
-"""for convenience"""		# TODO: contribute to ophyd?
-SCALER_AUTOCOUNT_MODE = 1
 
 
 def use_EPICS_scaler_channels(scaler):
@@ -149,9 +118,9 @@ class ApsMachineParametersDevice(Device):
     """
     common operational parameters of the APS of general interest
     
-    EXAMPLE::
+    USAGE::
 
-        import APS_BlueSky_tools.devices as APS_devices
+        import apstools.devices as APS_devices
         APS = APS_devices.ApsMachineParametersDevice(name="APS")
         aps_current = APS.current
 
@@ -209,11 +178,11 @@ class ApsMachineParametersDevice(Device):
         determine if APS is in User Operations mode (boolean)
         
         Use this property to configure ophyd Devices for direct or simulated hardware.
-        See issue #49 (https://github.com/BCDA-APS/APS_BlueSky_tools/issues/49) for details.
+        See issue #49 (https://github.com/BCDA-APS/apstools/issues/49) for details.
         
         EXAMPLE::
         
-            APS = APS_BlueSky_tools.devices.ApsMachineParametersDevice(name="APS")
+            APS = apstools.devices.ApsMachineParametersDevice(name="APS")
             
             if APS.inUserOperations:
                 suspend_APS_current = bluesky.suspenders.SuspendFloor(APS.current, 2, resume_thresh=10)
@@ -237,7 +206,7 @@ class ApsPssShutter(Device):
     * no indication that the shutter has actually moved from the bits
       (see :func:`ApsPssShutterWithStatus()` for alternative)
     
-    EXAMPLE::
+    USAGE::
     
         shutter_a = ApsPssShutter("2bma:A_shutter", name="shutter")
         
@@ -247,7 +216,9 @@ class ApsPssShutter(Device):
         shutter_a.set("open")
         shutter_a.set("close")
         
-    When using the shutter in a plan, be sure to use ``yield from``, such as::
+    When using the shutter in a plan, be sure to use ``yield from``.
+    
+    ::
 
         def in_a_plan(shutter):
             yield from abs_set(shutter, "open", wait=True)
@@ -321,14 +292,14 @@ class ApsPssShutter(Device):
 
 class ApsPssShutterWithStatus(Device):
     """
-    APS PSS shutter with separate status PV
+    APS PSS shutter
     
     * APS PSS shutters have separate bit PVs for open and close
     * set either bit, the shutter moves, and the bit resets a short time later
     * a separate status PV tells if the shutter is open or closed
       (see :func:`ApsPssShutter()` for alternative)
     
-    EXAMPLE::
+    USAGE::
     
         A_shutter = ApsPssShutterWithStatus(
             "2bma:A_shutter", 
@@ -434,7 +405,7 @@ class SimulatedApsPssShutterWithStatus(Device):
     """
     Simulated APS PSS shutter
     
-    EXAMPLE::
+    USAGE::
     
 		sim = SimulatedApsPssShutterWithStatus(name="sim")
     
@@ -526,9 +497,7 @@ class ApsUndulator(Device):
     """
     APS Undulator
     
-    EXAMPLE::
-    
-        undulator = ApsUndulator("ID09ds:", name="undulator")
+    USAGE:  ``undulator = ApsUndulator("ID09ds:", name="undulator")``
     """
     energy = Component(EpicsSignal, "Energy", write_pv="EnergySet")
     energy_taper = Component(EpicsSignal, "TaperEnergy", write_pv="TaperEnergySet")
@@ -557,9 +526,7 @@ class ApsUndulatorDual(Device):
     """
     APS Undulator with upstream *and* downstream controls
     
-    EXAMPLE::
-    
-        undulator = ApsUndulatorDual("ID09", name="undulator")
+    USAGE:  ``undulator = ApsUndulatorDual("ID09", name="undulator")``
     
     note:: the trailing ``:`` in the PV prefix should be omitted
     """
@@ -573,7 +540,7 @@ class ApsBssUserInfoDevice(Device):
     
     BSS: Beamtime Scheduling System
 
-    EXAMPLE::
+    USAGE::
 
         bss_user_info = ApsBssUserInfoDevice(
                             "9id_bss:",
@@ -600,29 +567,42 @@ class ApsBssUserInfoDevice(Device):
     esaf_team =         Component(EpicsSignal, "esaf_team",     string=True)
 
 
-class DeviceMixinBase(Device):
-    """Base class for APS_Bluesky_tools Device mixin classes"""
-
-
 class AxisTunerException(ValueError): 
     """Exception during execution of `AxisTunerBase` subclass"""
+    pass
 
 
 class AxisTunerMixin(EpicsMotor):
     """
     Mixin class to provide tuning capabilities for an axis
     
-    See the `TuneAxis()` example in this jupyter notebook: 
-    https://github.com/BCDA-APS/APS_BlueSky_tools/blob/master/docs/source/resources/demo_tuneaxis.ipynb
+    USAGE::
     
+        class TunableEpicsMotor(AxisTunerMixin, EpicsMotor):
+            pass
+        
+        def a2r_pretune_hook():
+            # set the counting time for *this* tune
+            yield from bps.abs_set(scaler.preset_time, 0.2)
+            
+        a2r = TunableEpicsMotor("xxx:m1", name="a2r")
+        a2r.tuner = TuneAxis([scaler], a2r, signal_name=scaler.channels.chan2.name)
+        a2r.tuner.width = 0.02
+        a2r.tuner.num = 21
+        a2r.pre_tune_method = a2r_pretune_hook
+        RE(a2r.tune())
+        
+        # tune four of the USAXS axes (using preconfigured parameters for each)
+        RE(tune_axes([mr, m2r, ar, a2r])
+
     HOOK METHODS
     
     There are two hook methods (`pre_tune_method()`, and `post_tune_method()`)
     for callers to add additional plan parts, such as opening or closing shutters, 
     setting detector parameters, or other actions.
     
-    Each hook method must accept a single argument: 
-    an axis object such as `EpicsMotor` or `SynAxis`,
+    Each hook method must accept one argument: 
+    axis object such as `EpicsMotor` or `SynAxis`,
     such as::
     
         def my_pre_tune_hook(axis):
@@ -630,7 +610,8 @@ class AxisTunerMixin(EpicsMotor):
         def my_post_tune_hook(axis):
             yield from bps.mv(shutter, "close")
         
-        class TunableSynAxis(AxisTunerMixin, SynAxis): pass
+        class TunableSynAxis(AxisTunerMixin, SynAxis):
+            pass
 
         myaxis = TunableSynAxis(name="myaxis")
         mydet = SynGauss('mydet', myaxis, 'myaxis', center=0.21, Imax=0.98e5, sigma=0.127)
@@ -642,9 +623,8 @@ class AxisTunerMixin(EpicsMotor):
 
     """
     
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.tuner = None   # such as: APS_BlueSky_tools.plans.TuneAxis
+    def __init__(self):
+        self.tuner = None   # such as: apstools.plans.TuneAxis
         
         # Hook functions for callers to add additional plan parts
         # Each must accept one argument: axis object such as `EpicsMotor` or `SynAxis`
@@ -662,7 +642,7 @@ class AxisTunerMixin(EpicsMotor):
     def tune(self, md=None, **kwargs):
         if self.tuner is None:
             msg = "Must define an axis tuner, none specified."
-            msg += "  Consider using APS_BlueSky_tools.plans.TuneAxis()"
+            msg += "  Consider using apstools.plans.TuneAxis()"
             raise AxisTunerException(msg)
         
         if self.tuner.axis is None:
@@ -684,63 +664,57 @@ class AxisTunerMixin(EpicsMotor):
                 self.post_tune_method()
 
 
-# TODO: issue #76
-# class TunableSynAxis(AxisTunerMixin, SynAxis): """synthetic axis that can be tuned"""
-# class TunableEpicsMotor(AxisTunerMixin, EpicsMotor): """EpicsMotor that can be tuned"""
-
-
-class EpicsDescriptionMixin(DeviceMixinBase):
+class TunableEpicsMotor(AxisTunerMixin, EpicsMotor):
     """
-    add a record's description field to a Device, such as EpicsMotor
+    EPICS motor with signal for tuning
     
-    EXAMPLE::
+    USAGE::
+
+        def a2r_pretune_hook():
+            # set the counting time for *this* tune
+            yield from bps.abs_set(scaler.preset_time, 0.2)
     
-        from ophyd import EpicsMotor
-        from APS_BlueSky_tools.devices import EpicsDescriptionMixin
+        a2r = TunableEpicsMotor("xxx:m1", name="a2r")
+        a2r.tuner = TuneAxis([scaler], a2r, signal_name=scaler.channels.chan2.name)
+        a2r.tuner.width = 0.02
+        a2r.tuner.num = 21
+        a2r.pre_tune_method = a2r_pretune_hook
+        RE(a2r.tune())
     
-        class myEpicsMotor(EpicsDescriptionMixin, EpicsMotor): pass
+    """
+    __metaclass__ = EpicsMotor
+
+
+class EpicsMotorDialMixin(object):
+    """
+    add motor record's dial coordinate fields
+    
+    USAGE::
+    
+        class myEpicsMotor(EpicsMotor, EpicsMotorDialMixin): pass
         m1 = myEpicsMotor('xxx:m1', name='m1')
-        print(m1.desc.value)
-    
-    """
-    
-    desc = Component(EpicsSignal, ".DESC")
-
-
-class EpicsMotorDialMixin(DeviceMixinBase):
-    """
-    add motor record's dial coordinate fields to Device
-    
-    EXAMPLE::
-    
-        from ophyd import EpicsMotor
-        from APS_BlueSky_tools.devices import EpicsMotorDialMixin
-
-        class myEpicsMotor(EpicsMotorDialMixin, EpicsMotor): pass
-        m1 = myEpicsMotor('xxx:m1', name='m1')
-        print(m1.dial.read())
     
     """
     
     dial = Component(EpicsSignal, ".DRBV", write_pv=".DVAL")
 
 
-class EpicsMotorLimitsMixin(DeviceMixinBase):
+class EpicsMotorWithDial(EpicsMotor, EpicsMotorDialMixin):
+    """
+    add motor record's dial coordinates to EpicsMotor
+    
+    USAGE::
+    
+        m1 = EpicsMotorWithDial('xxx:m1', name='m1')
+    
+    This is legacy support.  For new work, use `EpicsMotorDialMixin`.
+    """
+    pass
+
+
+class EpicsMotorLimitsMixin(Device):
     """
     add motor record HLM & LLM fields & compatibility get_lim() and set_lim()
-    
-    EXAMPLE::
-    
-        from ophyd import EpicsMotor
-        from APS_BlueSky_tools.devices import EpicsMotorLimitsMixin
-
-        class myEpicsMotor(EpicsMotorLimitsMixin, EpicsMotor): pass
-        m1 = myEpicsMotor('xxx:m1', name='m1')
-        lo = m1.get_lim(-1)
-        hi = m1.get_lim(1)
-        m1.set_lim(-25, -5)
-        print(m1.get_lim(-1), m1.get_lim(1))
-        m1.set_lim(lo, hi)
     """
     
     soft_limit_lo = Component(EpicsSignal, ".LLM")
@@ -750,92 +724,89 @@ class EpicsMotorLimitsMixin(DeviceMixinBase):
         """
         Returns the user limit of motor
         
-        * flag > 0: returns high limit
-        * flag < 0: returns low limit
-        * flag == 0: returns None
-        
-        Similar with SPEC command
+        flag > 0: returns high limit
+        flag < 0: returns low limit
+        flag == 0: returns None
         """
         if flag > 0:
-            return self.soft_limit_hi.value
+            return self.high_limit
         else:
-            return self.soft_limit_lo.value
+            return self.low_limit
     
     def set_lim(self, low, high):
         """
         Sets the low and high limits of motor
         
-        * No action taken if motor is moving.
         * Low limit is set to lesser of (low, high)
         * High limit is set to greater of (low, high)
-        
-        Similar with SPEC command
+        * No action taken if motor is moving. 
         """
         if not self.moving:
             self.soft_limit_lo.put(min(low, high))
             self.soft_limit_hi.put(max(low, high))
 
 
-class EpicsMotorServoMixin(DeviceMixinBase):
+class EpicsMotorServoMixin(object):
     """
-    add motor record's servo loop controls to Device
+    add motor record's servo loop controls
     
-    EXAMPLE::
+    USAGE::
     
-        from ophyd import EpicsMotor
-        from APS_BlueSky_tools.devices import EpicsMotorServoMixin
-
-        class myEpicsMotor(EpicsMotorServoMixin, EpicsMotor): pass
+        class myEpicsMotor(EpicsMotor, EpicsMotorServoMixin): pass
         m1 = myEpicsMotor('xxx:m1', name='m1')
-        print(m1.servo.read())
+    
     """
     
     # values: "Enable" or "Disable"
     servo = Component(EpicsSignal, ".CNEN", string=True)
 
 
-class EpicsMotorRawMixin(DeviceMixinBase):
+class EpicsMotorWithServo(EpicsMotor, EpicsMotorServoMixin):
     """
-    add motor record's raw coordinate fields to Device
+    extend basic motor support to enable/disable the servo loop controls
     
-    EXAMPLE::
+    USAGE::
     
-        from ophyd import EpicsMotor
-        from APS_BlueSky_tools.devices import EpicsMotorRawMixin
+        m1 = EpicsMotorWithDial('xxx:m1', name='m1')
     
-        class myEpicsMotor(EpicsMotorRawMixin, EpicsMotor): pass
+    This is legacy support.  For new work, use `EpicsMotorServoMixin`.
+    """
+    pass
+
+
+class EpicsMotorRawMixin(object):
+    """
+    add motor record's raw coordinate fields
+    
+    USAGE::
+    
+        class myEpicsMotor(EpicsMotor, EpicsMotorRawMixin): pass
         m1 = myEpicsMotor('xxx:m1', name='m1')
-        print(m1.raw.read())
+    
     """
     
     raw = Component(EpicsSignal, ".RRBV", write_pv=".RVAL")
 
 
-# TODO: issue #76
-# class EpicsMotorWithDescription(EpicsDescriptionMixin, EpicsMotor):
-#     """EpicsMotor with description field"""
-#
-# class EpicsMotorWithMore(
-#     EpicsDescriptionMixin, 
-#     EpicsMotorLimitsMixin, 
-#     EpicsMotorDialMixin,
-#     EpicsMotorRawMixin, 
-#     EpicsMotor): 
-#     """
-#     EpicsMotor with more fields
-#     
-#     * description (``desc``)
-#     * soft motor limits (``soft_limit_hi``, ``soft_limit_lo``)
-#     * dial coordinates (``dial``)
-#     * raw coordinates (``raw``)
-#     """
+class EpicsMotorDescriptionMixin(object):
+    """
+    add motor record's description field
+    
+    USAGE::
+    
+        class myEpicsMotor(EpicsMotor, EpicsMotorDescriptionMixin): pass
+        m1 = myEpicsMotor('xxx:m1', name='m1')
+    
+    """
+    
+    desc = Component(EpicsSignal, ".DESC")
 
 
 class EpicsMotorShutter(Device):
     """
     a shutter, implemented with an EPICS motor moved between two positions
     
-    EXAMPLE::
+    USAGE::
 
         tomo_shutter = EpicsMotorShutter("2bma:m23", name="tomo_shutter")
         tomo_shutter.closed_position = 1.0      # default
@@ -932,7 +903,7 @@ class EpicsOnOffShutter(Device):
     The current position is determined by comparing the value of the control
     with the expected open and close values.
     
-    EXAMPLE::
+    USAGE::
 
         bit_shutter = EpicsOnOffShutter("2bma:bit1", name="bit_shutter")
         bit_shutter.closed_position = 0      # default
@@ -1015,7 +986,7 @@ class DualPf4FilterBox(Device):
     """
     Dual Xia PF4 filter boxes using support from synApps (using Al, Ti foils)
     
-    EXAMPLE::
+    Example::
     
         pf4 = DualPf4FilterBox("2bmb:pf4:", name="pf4")
         pf4_AlTi = DualPf4FilterBox("9idcRIO:pf4:", name="pf4_AlTi")
@@ -1248,31 +1219,15 @@ def AD_warmed_up(detector):
     return verdict
 
 
-class AD_EpicsHdf5FileName(FileStorePluginBase):
+class ApsFileStoreHDF5(FileStorePluginBase):
     """
     custom class to define image file name from EPICS
-    
-    .. caution:: *Caveat emptor* applies here.  You assume expertise!
-    
-    Replace standard Bluesky algorithm where file names
-    are defined as UUID strings, virtually guaranteeing that 
-    no existing images files will ever be overwritten.
-    
-    Also, this method decouples the data files from the databroker,
-    which needs the files to be named by UUID.
-    
-    .. autosummary::
-       
-        ~make_filename
-        ~generate_datum
-        ~get_frames_per_point
-        ~stage
 
-    To allow users to control the file **name**,
+    To allow users to control the file name,
     we override the ``make_filename()`` method here
     and we need to override some intervening classes.
 
-    To allow users to control the file **number**,
+    To allow users to control the file number,
     we override the ``stage()`` method here
     and triple-comment out that line, and bring in
     sections from the methods we are replacing here.
@@ -1280,75 +1235,17 @@ class AD_EpicsHdf5FileName(FileStorePluginBase):
     The image file name is set in `FileStoreBase.make_filename()` 
     from `ophyd.areadetector.filestore_mixins`.  This is called 
     (during device staging) from `FileStoreBase.stage()`
-    
-    EXAMPLE:
 
     To use this custom class, we need to connect it to some
-    intervening structure.  Here are the steps:
-    
-    #. override default file naming
-    #. use to make your custom iterative writer
-    #. use to make your custom HDF5 plugin
-    #. use to make your custom AD support
-    
-    imports::
+    intervening structure:
 
-        from bluesky import RunEngine, plans as bp
-        from ophyd.areadetector import SimDetector, SingleTrigger
-        from ophyd.areadetector import ADComponent, ImagePlugin, SimDetectorCam
-        from ophyd.areadetector import HDF5Plugin
-        from ophyd.areadetector.filestore_mixins import FileStoreIterativeWrite
-    
-    override default file naming::
-        
-        from APS_BlueSky_tools.devices import AD_EpicsHdf5FileName
-    
-    make a custom iterative writer::
-        
-        class myHdf5EpicsIterativeWriter(AD_EpicsHdf5FileName, FileStoreIterativeWrite): pass
-    
-    make a custom HDF5 plugin::
-        
-        class myHDF5FileNames(HDF5Plugin, myHdf5EpicsIterativeWriter): pass
-    
-    define support for the detector (simulated detector here)::
-        
-        class MySimDetector(SingleTrigger, SimDetector):
-            '''SimDetector with HDF5 file names specified by EPICS'''
-            
-            cam = ADComponent(SimDetectorCam, "cam1:")
-            image = ADComponent(ImagePlugin, "image1:")
-            
-            hdf1 = ADComponent(
-                myHDF5FileNames, 
-                suffix = "HDF1:", 
-                root = "/",
-                write_path_template = "/",
-                )
-    
-    create an instance of the detector::
-        
-        simdet = MySimDetector("13SIM1:", name="simdet")
-        if hasattr(simdet.hdf1.stage_sigs, "array_counter"):
-            # remove this so array counter is not set to zero each staging
-            del simdet.hdf1.stage_sigs["array_counter"]
-        simdet.hdf1.stage_sigs["file_template"] = '%s%s_%3.3d.h5'
-    
-    setup the file names using the EPICS HDF5 plugin::
-        
-        simdet.hdf1.file_path.put("/tmp/simdet_demo/")    # ! ALWAYS end with a "/" !
-        simdet.hdf1.file_name.put("test")
-        simdet.hdf1.array_counter.put(0)
-    
-    If you have not already, create a bluesky RunEngine::
-        
-        RE = RunEngine({})
-    
-    take an image::
-
-        RE(bp.count([simdet]))
-    
-    INTERNAL METHODS
+    ====================================  ============================
+    custom class                          superclass(es)
+    ====================================  ============================
+    ``ApsFileStoreHDF5``                  ``FileStorePluginBase``
+    ``ApsFileStoreHDF5IterativeWrite``    ``ApsFileStoreHDF5``, `FileStoreIterativeWrite``
+    ``ApsHDF5Plugin``                     ``HDF5Plugin``, `ApsFileStoreHDF5IterativeWrite``
+    ====================================  ============================
     """
 
     def __init__(self, *args, **kwargs):
@@ -1391,16 +1288,11 @@ class AD_EpicsHdf5FileName(FileStorePluginBase):
         return super().generate_datum(key, timestamp, datum_kwargs)
 
     def get_frames_per_point(self):
-        """overrides default behavior"""
+        """ """
         return self.num_capture.get()
 
     def stage(self):
-        """
-        overrides default behavior
-        
-        Set EPICS items before device is staged, then copy EPICS 
-        naming template (and other items) to ophyd after staging.
-        """
+        """ """
         # Make a filename.
         filename, read_path, write_path = self.make_filename()
 
@@ -1437,3 +1329,47 @@ class AD_EpicsHdf5FileName(FileStorePluginBase):
         # from FileStoreHDF5.stage()
         res_kwargs = {'frame_per_point': self.get_frames_per_point()}
         self._generate_resource(res_kwargs)
+
+
+class ApsFileStoreHDF5IterativeWrite(ApsFileStoreHDF5, FileStoreIterativeWrite):
+    """custom class to enable users to control image file name"""
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        FileStoreIterativeWrite.__init__(self, *args, **kwargs)
+
+
+class ApsHDF5Plugin(HDF5Plugin, ApsFileStoreHDF5IterativeWrite):
+    """
+    custom class to take image file names from EPICS
+    
+    NOTE: replaces standard Bluesky algorithm where file names
+          are defined as UUID strings, virtually guaranteeing that 
+          no existing images files will ever be overwritten.
+          *Caveat emptor* applies here.  You assume some expertise!
+    
+    USAGE::
+
+        class MySimDetector(SingleTrigger, SimDetector):
+            '''SimDetector with HDF5 file names specified by EPICS'''
+            
+            cam = ADComponent(MyAltaCam, "cam1:")
+            image = ADComponent(ImagePlugin, "image1:")
+            
+            hdf1 = ADComponent(
+                ApsHDF5Plugin, 
+                suffix = "HDF1:", 
+                root = "/",
+                write_path_template = "/local/data",
+                )
+
+        simdet = MySimDetector("13SIM1:", name="simdet")
+        # remove this so array counter is not set to zero each staging
+        del simdet.hdf1.stage_sigs["array_counter"]
+        simdet.hdf1.stage_sigs["file_template"] = '%s%s_%3.3d.h5'
+        simdet.hdf1.file_path.put("/local/data/demo/")
+        simdet.hdf1.file_name.put("test")
+        simdet.hdf1.array_counter.put(0)
+        RE(bp.count([simdet]))
+
+    """
