+++ docs/source/applications.rst
@@ -2,12 +2,12 @@
 Applications
 ------------
 
-There are two command-line applications provided by APS_BlueSky_tools:
+There are two command-line applications provided by apstools:
 
-=============================================================   =================================
+====================================================   =================================
 application   purpose
-=============================================================   =================================
-:ref:`aps_bluesky_tools_plan_catalog <example_plan_catalog>`    summary list of all scans in the databroker 
-:ref:`bluesky_snapshot`                                         Take a snapshot of a list of EPICS PVs and record it in the databroker.
-=============================================================   =================================
+====================================================   =================================
+:ref:`apstools_plan_catalog <example_plan_catalog>`    summary list of all scans in the databroker 
+:ref:`bluesky_snapshot`                                Take a snapshot of a list of EPICS PVs and record it in the databroker.
+====================================================   =================================
