+++ docs/source/index.rst
@@ -1,10 +1,10 @@
-.. APS_BlueSky_tools documentation master file, created by
+.. apstools documentation master file, created by
    sphinx-quickstart on Wed Nov 15 14:58:09 2017.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to APS_BlueSky_tools's documentation!
-=============================================
+Welcome to apstools's documentation!
+====================================
 
 Various Python tools for use with BlueSky at the APS
 
@@ -18,8 +18,8 @@ Package Information
 :email:         jemian@anl.gov
 :copyright:     2017-2019, Pete R. Jemian
 :license:       ANL OPEN SOURCE LICENSE (see LICENSE file)
-:documentation: https://APS_BlueSky_tools.readthedocs.io
-:source:        https://github.com/BCDA-APS/APS_BlueSky_tools
+:documentation: https://apstools.readthedocs.io
+:source:        https://github.com/BCDA-APS/apstools
 
 .. toctree::
    :maxdepth: 2
