+++ docs/source/source/signals.rst
@@ -2,6 +2,6 @@
 Signals
 -------
 
-.. automodule:: APS_BlueSky_tools.signals
+.. automodule:: apstools.signals
     :members:
