+++ docs/source/source/suspenders.rst
@@ -2,6 +2,6 @@
 Suspenders
 ----------
 
-.. automodule:: APS_BlueSky_tools.suspenders
+.. automodule:: apstools.suspenders
     :members:
