+++ docs/source/synApps/busy.rst
@@ -5,6 +5,6 @@ synApps busy record
 see the synApps ``busy`` module suppport: 
 https://github.com/epics-modules/busy
 
-.. automodule:: APS_BlueSky_tools.synApps_ophyd.busy
+.. automodule:: apstools.synApps_ophyd.busy
     :members:
