+++ docs/source/synApps/sscan.rst
@@ -5,6 +5,6 @@ synApps sscan record
 see the synApps ``sscan`` module suppport: 
 https://github.com/epics-modules/sscan
 
-.. automodule:: APS_BlueSky_tools.synApps_ophyd.sscan
+.. automodule:: apstools.synApps_ophyd.sscan
     :members:
