+++ docs/source/synApps/swait.rst
@@ -8,6 +8,6 @@ https://htmlpreview.github.io/?https://raw.githubusercontent.com/epics-modules/c
 see the synApps ``calc`` module suppport: 
 https://github.com/epics-modules/calc
 
-.. automodule:: APS_BlueSky_tools.synApps_ophyd.swait
+.. automodule:: apstools.synApps_ophyd.swait
     :members:
