+++ setup.cfg
@@ -11,7 +11,7 @@ release = egg_info -RDb ''
 [versioneer]
 VCS = git
 style = pep440
-versionfile_source = APS_BlueSky_tools/_version.py
-versionfile_build = APS_BlueSky_tools/_version.py
+versionfile_source = apstools/_version.py
+versionfile_build = apstools/_version.py
 tag_prefix =
-parentdir_prefix = APS_BlueSky_tools- 
+parentdir_prefix = apstools-
